Fix version number issue
Hi,  I made a mistake and released a version 1.3.0 instead of 0.3.0.  We've deleted the release and released it with the correct version number, but dbt hub picked up 1.3.0.  I opened a pr and @dbeatty10 helped resolve the issue [here](https://github.com/dbt-labs/hubcap/pull/254) , but it seems like the problem has reverted... It looks like hubcap updated dbt hub again with the deleted release.  I'm not sure if hubcap is picking up the deleted version number, or if there is some vestige of the old deleted release persisting in the hubcap process somewhere.

Let me know what I can do to help fix, thanks and apologies again.